requirements.txt
# Django Framework
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.61.0
dj-rest-auth==5.0.2

# WebSocket + ASGI
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Cache / Broker
redis==5.0.1

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
lxml>=4.9.0
yfinance==0.2.36
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1

# Static files
whitenoise==6.6.0

# Admin UI
django-jazzmin==2.6.0

# Production server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
